Updates integration tests to latests updates
